Level Design
This update adds more enemies on the gird, also, James was able to spawn
terrain, which you will see when ran. The terrain has no collision
properties yet so until we figure that out the player just goes right
through them, but the grid is now populated with objects. I think we
might have to change behavior of enemies when we do manage to get the
collisions sorted out, given that I am not sure what will happen if they
run into an obstical blocking them from the player. Also you will notice
more grids floating above the main grid, when we figure out the
Jumping/Jetpack addition to the player, these grids will be explorable.